# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Contract Hub API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1"
        }
    ],
    "paths": {
        "/counterparties": {
            "get": {
                "tags": [
                    "Counterparties"
                ],
                "summary": "Получить контрагентов",
                "operationId": "d77b24f60572df2d061c74f7cc47f4ef",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список контрагентов"
                    },
                    "401": {
                        "description": "Неавторизован"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/counterparty": {
            "post": {
                "tags": [
                    "Counterparties"
                ],
                "summary": "Добавить контрагента",
                "operationId": "d4c4ebfcb9cc1cfe66884e0d6585ce43",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "inn"
                                ],
                                "properties": {
                                    "inn": {
                                        "type": "string",
                                        "example": "7736207543"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Контрагент добавлен"
                    },
                    "401": {
                        "description": "Неавторизован"
                    },
                    "404": {
                        "description": "Контрагент не найден"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Авторизация пользователя",
                "operationId": "9a5afa5942b05962cb3c74ed5e7b197c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "1@1.ru"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная авторизация"
                    },
                    "401": {
                        "description": "Неверные данные"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Выход из системы",
                "operationId": "8fc6ca86230798a6d20d84f6d3c48ea1",
                "responses": {
                    "200": {
                        "description": "Успешный выход"
                    },
                    "401": {
                        "description": "Неавторизован"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/registration": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Регистрация пользователя",
                "operationId": "e21799a4e4234235d603e7c1282e6e7e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "lastname": {
                                        "type": "string",
                                        "example": "Иванов",
                                        "nullable": true
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Иван",
                                        "nullable": true
                                    },
                                    "patronymic": {
                                        "type": "string",
                                        "example": "Иванович",
                                        "nullable": true
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "1@1.ru"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Успешная регистрация"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "http",
                "description": "Введите токен (без \"Bearer\"). Получите токен через POST /api/v1/login",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "tags": [
        {
            "name": "Counterparties",
            "description": "Counterparties"
        },
        {
            "name": "Authentication",
            "description": "Authentication"
        }
    ],
    "security": [
        {
            "sanctum": []
        }
    ]
}